pyproject.toml
[project]
name = "v2"
version = "0.1.0"
description = "SQL Agent for High-Cardinality Databases"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.8",
    "chromadb>=1.1.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.1.1",
]
